Move test-runner.js to tests directory.
I also moved dom-shim-tests.js to its own directory for
better organization.
